requirements.txt
httpx>=0.27.0
parsel>=1.9.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
